+++ python/private/stage1_bootstrap_template.sh
@@ -16,7 +16,9 @@ PYTHON_BINARY='%python_binary%'
 IS_ZIPFILE="%is_zipfile%"
 
 if [[ "$IS_ZIPFILE" == "1" ]]; then
-  zip_dir=$(mktemp -d --suffix Bazel.runfiles_)
+  # NOTE: Macs have an old version of mktemp, so we must use only the
+  # minimal functionality of it.
+  zip_dir=$(mktemp -d)
 
   if [[ -n "$zip_dir" && -z "${RULES_PYTHON_BOOTSTRAP_VERBOSE:-}" ]]; then
     trap 'rm -fr "$zip_dir"' EXIT
@@ -27,7 +29,7 @@ if [[ "$IS_ZIPFILE" == "1" ]]; then
   # The alternative requires having to copy ourselves elsewhere with the prelude
   # stripped (because zip can't extract from a stream). We avoid that because
   # it's wasteful.
-  ( unzip -q -d "$zip_dir" "$0" 2>/dev/null || /bin/true )
+  ( unzip -q -d "$zip_dir" "$0" 2>/dev/null || true )
 
   RUNFILES_DIR="$zip_dir/runfiles"
   if [[ ! -d "$RUNFILES_DIR" ]]; then
@@ -105,14 +107,34 @@ declare -a interpreter_args
 # NOTE: Only works for 3.11+
 interpreter_env+=("PYTHONSAFEPATH=1")
 
-export RUNFILES_DIR
-# NOTE: We use <(...) to pass the Python program as a file so that stdin can
-# still be passed along as normal.
-env \
-  "${interpreter_env[@]}" \
-  "$python_exe" \
-  "${interpreter_args[@]}" \
-  "$stage2_bootstrap" \
-  "$@"
+if [[ "$IS_ZIPFILE" == "1" ]]; then
+  interpreter_args+=("-XRULES_PYTHON_ZIP_DIR=$zip_dir")
+fi
 
-exit $?
+
+export RUNFILES_DIR
+
+command=(
+  env
+  "${interpreter_env[@]}"
+  "$python_exe"
+  "${interpreter_args[@]}"
+  "$stage2_bootstrap"
+  "$@"
+)
+
+# We use `exec` instead of a child process so that signals sent directly (e.g.
+# using `kill`) to this process (the PID seen by the calling process) are
+# received by the Python process. Otherwise, this process receives the signal
+# and would have to manually propagate it.
+# See https://github.com/bazelbuild/rules_python/issues/2043#issuecomment-2215469971
+# for more information.
+#
+# However, when running a zip file, we need to clean up the workspace after the
+# process finishes so control must return here.
+if [[ "$IS_ZIPFILE" == "1" ]]; then
+  "${command[@]}"
+  exit $?
+else
+  exec "${command[@]}"
+fi